devide parser to two files
