+++ .gitignore
@@ -32,3 +32,4 @@ Thumbs.db
 # מודול תלויות (אם רוצים לבנות מחדש עם go mod tidy)
 # אבל לרוב משאירים אותו בגיט
 # /vendor/
+credentials.json
