+++ Parser/parse_json.go
@@ -3,12 +3,31 @@ package Parser
 import (
 	"encoding/json"
 	"fmt"
-	"strconv"
 	"time"
 
 	"github.com/HadasAmar/analytics-load-tool/Model"
 	"github.com/HadasAmar/analytics-load-tool/Reader"
 )
+
+type LogEntry = Model.LogEntry
+
+func ParseRecord(r Reader.RawRecord) *LogEntry {
+	var query Model.QueryData
+	fmt.Println("📥 Raw JSON:", r.RawQuery)
+	err := json.Unmarshal([]byte(r.RawQuery), &query)
+	if err != nil {
+		fmt.Println("Error parsing query:", err)
+		return nil
+	}
+	t, err := time.Parse(time.RFC3339, r.Timestamp)
+	if err != nil {
+		fmt.Println("Error parsing timestamp:", err)
+		return nil
+	}
+	fmt.Println("Parsed query:", query)
+	entry := FromQueryRecord(t, r.IP, query)
+	return &entry
+}
 
 func FromQueryRecord(timestamp time.Time, ip string, query Model.QueryData) Model.LogEntry {
 	fmt.Println("Parsing query record:", query)
@@ -40,55 +59,6 @@ func FromQueryRecord(timestamp time.Time, ip string, query Model.QueryData) Mode
 	}
 }
 
-func ParseCSVRaw(r Reader.RawRecord) *Model.LogEntry {
-	var row map[string]string
-	if err := json.Unmarshal([]byte(r.RawQuery), &row); err != nil {
-		return nil
-	}
-	entry := Model.LogEntry{
-		CampaignID:          row["campaign_id"],
-		Partner:             row["partner"],
-		AppID:               row["app_id"],
-		UnmaskedMediaSource: row["unmasked_media_source"],
-		MediaSource:         row["media_source"],
-		AttributionType:     row["attribution_type"],
-		Campaign:            row["campaign"],
-		Source:              row["source"],
-		AdID:                row["ad_id"],
-		AdsetID:             row["adset_id"],
-		AdsetName:           row["adset_name"],
-		SiteID:              row["site_id"],
-		Ad:                  row["ad"],
-		LtvCountry:          row["ltv_country"],
-		Installs:            parseInt(row["installs"]),
-		Impressions:         parseInt(row["impressions"]),
-		Clicks:              parseInt(row["clicks"]),
-		Loyals:              parseInt(row["loyals"]),
-		OrganicInstalls:     parseInt(row["organic_installs"]),
-		OrganicImpressions:  parseInt(row["organic_impressions"]),
-		OrganicClicks:       parseInt(row["organic_clicks"]),
-		OrganicLoyals:       parseInt(row["organic_loyals"]),
-		LogTime:             parseTime(row["log_time"]),
-		IP:                  row["ip"],
-	}
-	return &entry
-}
-
-func parseInt(s string) int {
-	i, err := strconv.Atoi(s)
-	if err != nil {
-		return 0
-	}
-	return i
-}
-
-func parseTime(s string) time.Time {
-	t, err := time.Parse(time.RFC3339, s)
-	if err != nil {
-		return time.Time{}
-	}
-	return t
-}
 
 func GetDim(dims []Model.DimensionSpec, name string) string {
 	for _, d := range dims {
@@ -107,3 +77,4 @@ func GetAgg(aggs []Model.AggregationSpec, name string) int {
 	}
 	return 0
 }
+
